+++ sample-app/src/main/java/com/segment/analytics/kotlin/brazesample/testapp/MainActivity.kt
@@ -6,6 +6,7 @@ import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import com.braze.support.BrazeLogger
 import com.segment.analytics.kotlin.brazesample.testapp.databinding.ActivityMainBinding
+import kotlinx.serialization.json.*
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
@@ -20,7 +21,30 @@ class MainActivity : AppCompatActivity() {
         }
 
         findViewById<Button>(R.id.identifyButton).setOnClickListener {
-            MainApplication.analytics.identify("bob")
+            MainApplication.analytics.identify("bob", buildJsonObject {
+                put("username", "BobBraze")
+                put("email", "bob@test.com")
+                put("plan", "premium")
+                put("testArray", buildJsonArray {
+                    add("test")
+                    add(3)
+                    add(true)
+                }
+                )
+                put("jobInfo", buildJsonObject {
+                    put("department", "G9D")
+                    put("office", "030-2 E208")
+                    put("skillInfo", buildJsonObject {
+                        put("certified", true)
+                        put("languages", buildJsonArray {
+                            add("Kotlin")
+                            add("Java")
+                            add(1)
+                            add(false)
+                        })
+                    })
+                })
+            });
         }
     }
 }